aider: Added form and mutation to create a Sphere when the user has filled out the form correctly.
# Aider chat conversation:

USER: In the CreateSphere component, use the createSphere graphql mutation and a new form to add a Sphere when the user has filled out the form correctly.

ASSISTANT: To achieve this, we need to import the `useMutation` hook from `@apollo/client` and the `createSphere` mutation from your GraphQL mutations. We also need to create a form for the user to fill out.

Here are the changes:

```tsx
packages/app/src/components/CreateSphere.tsx
<<<<<<< SEARCH
import React from 'react';

function CreateSphere() {
  return (
    <div>
      <h2>Create Sphere</h2>
      {/* Add your form or other UI elements here */}
    </div>
  );
}

export default CreateSphere;
=======
import React, { useState } from 'react';
import { useMutation } from '@apollo/client';
import { CREATE_SPHERE } from '../graphql/mutations/sphere';

function CreateSphere() {
  const [name, setName] = useState('');
  const [createSphere, { data }] = useMutation(CREATE_SPHERE);

  const handleSubmit = async (event) => {
    event.preventDefault();
    try {
      await createSphere({ variables: { name } });
    } catch (error) {
      console.error(error);
    }
  };

  return (
    <div>
      <h2>Create Sphere</h2>
      <form onSubmit={handleSubmit}>
        <label>
          Name:
          <input type="text" value={name} onChange={(e) => setName(e.target.value)} required />
        </label>
        <button type="submit">Create Sphere</button>
      </form>
    </div>
  );
}

export default CreateSphere;
>>>>>>> REPLACE
```

Please replace `CREATE_SPHERE` with the actual import path of your `createSphere` mutation. The form includes a single input field for the name of the sphere. When the form is submitted, it calls the `createSphere` mutation with the name as a variable. If the mutation is successful, the new sphere will be created. If there is an error, it will be logged to the console.