aider: Update the `CREATE_ORBIT` mutation in `orbits.ts` to match the schema and return base64 hashes for the `actionHash` and `entryHash` values.
# Aider chat conversation:

USER: Update the orbits.ts mocks so that the CREATE_ORBIT mutation has variables that match the schema. Change the returned headerHash to be an actionHash and the values of the payload to actually be b64 hashes.

ASSISTANT: To update the `CREATE_ORBIT` mutation in the `orbits.ts` file, we need to modify the variables to match the schema and change the returned `headerHash` to `actionHash`. Also, we need to update the values of the payload to be base64 hashes.

Here is the *SEARCH/REPLACE block* for the required changes:

```typescript
packages/app/src/graphql/mocks/orbits.ts
<<<<<<< SEARCH
{
  request: {
    query: CREATE_ORBIT,
    variables: {
      variables:{
      name: "ABC",
      hashtag: "ABC",
      description: "ABC",
    }},
  },
  result: {
    data: {
      createSphere: {
        payload: {
          headerHash: "mockedHeaderHash",
          entryHash: "mockedEntryHash",
        },
      },
    },
  },
}
=======
{
  request: {
    query: CREATE_ORBIT,
    variables: {
      name: "ABC",
      hashtag: "ABC",
      description: "ABC",
    },
  },
  result: {
    data: {
      createOrbit: {
        actionHash: "bW9ja2VkQWN0aW9uSGFzaA==", // Base64 for "mockedActionHash"
        entryHash: "bW9ja2VkRW50cnlIYXNo", // Base64 for "mockedEntryHash"
      },
    },
  },
}
>>>>>>> REPLACE
```

Please note that the base64 values are just examples and should be replaced with actual base64 hashes.