aider: Added query and rendered SphereCard component with fetched sphere data in ListOrbits.
# Aider chat conversation:

USER: At the top of each `packages/app/src/components/lists/ListOrbits.tsx` I need to add a `packages/design-system/cards/SphereCard.tsx` that has the details of the Sphere with the id that was given to ListSpheres as a prop. This means I will need to do a getSphere query and give the fetched Sphere to the card as props. Make the changes to add the query, imported from the queries/sphere/GET_SPHERE.graphql file, and add the card after the `packages/app/src/components/lists/ListSortFilter.tsx` component but before the keyed list of orbits.

ASSISTANT: To achieve the requested functionality, we need to modify the `ListOrbits.tsx` component to include a query for fetching a specific sphere's details and then render the `SphereCard` component with the fetched data. We will use the `GET_SPHERE` query, which should be imported from the `queries/sphere/getSphere.graphql` file.

Here are the steps we need to take:

1. Import the `GET_SPHERE` query.
2. Add a new query hook to fetch the sphere details using the provided `sphereId`.
3. Render the `SphereCard` component with the fetched sphere data.

Let's start by modifying the `ListOrbits.tsx` file:

```typescript
packages/app/src/components/lists/ListOrbits.tsx
<<<<<<< SEARCH
import { useQuery } from '@apollo/client';
import { OrbitEdge } from '../../graphql/mocks/generated';
=======
import { useQuery, useLazyQuery } from '@apollo/client';
import { OrbitEdge, Sphere } from '../../graphql/mocks/generated';
import GET_SPHERE from '../../graphql/queries/sphere/getSphere.graphql';
>>>>>>> REPLACE

...

<<<<<<< SEARCH
  const { loading, error, data } = useQuery(GET_ORBITS_BY_SPHERE, {
=======
  const { loading: loadingOrbits, error: errorOrbits, data: dataOrbits } = useQuery(GET_ORBITS_BY_SPHERE, {
>>>>>>> REPLACE

...

<<<<<<< SEARCH
  if (loading) return <p>Loading...</p>;
  if (error) return <p>Error : {error.message}</p>;
=======
  const [getSphere, { loading: loadingSphere, data: dataSphere }] = useLazyQuery<Sphere>(GET_SPHERE, {
    variables: { id: sphereId },
  });

  // Fetch sphere details when component mounts if sphereId is provided
  useEffect(() => {
    if (sphereId) {
      getSphere();
    }
  }, [sphereId, getSphere]);

  if (loadingOrbits || loadingSphere) return <p>Loading...</p>;
  if (errorOrbits) return <p>Error : {errorOrbits.message}</p>;
>>>>>>> REPLACE

...

<<<<<<< SEARCH
      <ListSortFilter />
      <div className="orbits-list">
=======
      <ListSortFilter />
      {dataSphere && <SphereCard sphere={dataSphere.sphere} />}
      <div className="orbits-list">
>>>>>>> REPLACE
```

Please note that the `SphereCard` component expects a `sphere` prop of type `Sphere`, which should match the data structure returned by the `GET_SPHERE` query. Also, we've changed the `loading` and `error` variables to `loadingOrbits` and `errorOrbits` to differentiate them from the new `loadingSphere` and `errorSphere` variables for the sphere query.

Make sure that the `GET_SPHERE` query is correctly defined in the `queries/sphere/getSphere.graphql` file and that it returns the expected data structure for a `Sphere`.

If you have the `getSphere.graphql` file ready, please add it to the chat so I can include the correct import statement and any other necessary adjustments. If you need to create this file, let me know, and I can provide the content for it as well.